market-intel-brain-main/requirements_production.txt
# Production Dependencies - LIVE_PRODUCTION Mode
# Optimized for 8GB RAM + Mechanical HDD

# Core async framework
asyncio-mqtt==0.16.1
websockets==12.0
aiohttp==3.9.1
aiofiles==23.2.1

# Security & Encryption
cryptography==41.0.8
pycryptodome==3.20.0

# Data Processing & Validation
pydantic==2.12.5
python-dateutil==2.8.2
decimal==1.70

# Compression (HDD Optimization)
lz4==4.3.2
python-snappy==0.7.2

# HTTP Clients
httpx==0.28.1
requests==2.31.0

# JSON & Serialization
orjson==3.10.7
msgpack==1.0.8

# Monitoring & Performance
psutil==5.9.8
prometheus-client==0.19.0

# Database (if needed)
asyncpg==0.29.0
aioredis==2.0.1

# Logging
structlog==23.2.0
colorlog==6.8.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Type Checking
mypy==1.8.0
types-requests==2.31.0.20240125
